Custom delete exceptions covering DbIntegrity and NotFound exceptions

+++ src/main/java/stefano/course/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import stefano.course.entities.User;
 import stefano.course.repositories.UserRepository;
+import stefano.course.services.exceptions.DatabaseException;
 import stefano.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -30,7 +32,18 @@ public class UserService {
     }
 
     public void delete(Long id){
-        userRepository.deleteById(id);
+        try{
+            if(!userRepository.existsById(id)){
+                throw new ResourceNotFoundException(id);
+            }
+            userRepository.deleteById(id);
+        }
+        catch(ResourceNotFoundException e){
+            throw new ResourceNotFoundException(id);
+        }
+        catch(DataIntegrityViolationException e){
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj){
